Ejercicio_3/requirements.txt
# Ejercicio 3 - Sistema RAG + Agente Critico
# Python 3.11+

# Data processing
pandas==2.1.4
numpy==1.26.2

# YAML configuration
PyYAML==6.0.1

# Environment variables
python-dotenv==1.0.0

# LLM API (Groq)
groq==0.4.2

# Embeddings y Retrieval
sentence-transformers==2.2.2
torch>=2.0.0

# Token counting
tiktoken==0.5.2

# Evaluation y metricas
scikit-learn==1.3.2

# Hugging Face (dependencia de sentence-transformers)
transformers>=4.41.0
huggingface-hub>=0.20.0
